feat: improved custom block-breaking system for 1.20.5+ servers
fixup! feat: improved custom block-breaking system for 1.20.5+ servers

+++ core/src/main/java/io/th0rgal/oraxen/mechanics/provided/gameplay/BreakableMechanic.java
@@ -18,7 +18,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 public class BreakableMechanic {
 
-    private final int hardness;
+    private final double hardness;
     private final Drop drop;
     private final String itemId;
 
@@ -33,7 +33,7 @@ public class BreakableMechanic {
         return drop;
     }
 
-    public int hardness() {
+    public double hardness() {
         return hardness;
     }
